+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -14,6 +14,10 @@ Unreleased
 Changed
 *******
 - Fix GameSessionService.PeerConnectionRejected
+Added
+*****
+- Rework of server pools
+- Add back game server support in gamesession.
 
 5.3.0
 -----